% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_update_mvn.R
\name{conjugate_update_mvn_REH}
\alias{conjugate_update_mvn_REH}
\title{Conjugate update mu0 in the REH example}
\usage{
conjugate_update_mvn_REH(data, pars)
}
\arguments{
\item{data}{List with \verb{G, P}}

\item{pars}{List with \code{phi} and \code{theta}}
}
\description{
Conjugate update mu0 in the REH example
}
