% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter counts based on number of zeros and mean counts, per gene.}
\usage{
filter_counts(counts, group, options)
}
\arguments{
\item{counts}{Data frame of counts with column \code{Geneid}}

\item{group}{Vector of group labels for each sample}

\item{options}{List with some of \code{n0}, \code{n0_grp}, \code{mean_count},
\code{mean_count_grp};}
}
\description{
Filter counts based on number of zeros and mean counts, per gene.
}
