% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ribovi.R
\name{fit_ncvi}
\alias{fit_ncvi}
\title{Non-conjugate variational inference}
\usage{
fit_ncvi(
  data,
  init,
  update_pars,
  elbo,
  options = list(max_iter = 100, elbo_delta = 1e-04, verbose = TRUE, fixed_iter = FALSE,
    short_out = FALSE),
  ...
)
}
\arguments{
\item{data}{List of observed or known variables used by the
differentials or update functions}

\item{init}{List of initial values for unknown variables
updated by the algorithm}

\item{update_pars}{Function to carry out the update to
\code{pars} at each step}

\item{elbo}{Function to calculate the ELBO for the model}

\item{options}{List of options.
Should include positive real 'elbo_delta', the threshold
change in ELBO to terminate the algorithm}

\item{...}{Additional arguments to be passed to \code{update_pars()},
e.g. \code{differentials}}
}
\description{
General form of wrapper function for a
variational inference algorithm that may include
both conjugate and non-conjugate conditional distributions.
}
