% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_update_mvn.R
\name{nc_update_mvn}
\alias{nc_update_mvn}
\title{Non-conjugate update for the parameters of
multivariate normal distribution.}
\usage{
nc_update_mvn(data, pars, differentials, i)
}
\arguments{
\item{data}{List of observed and known variables passed down from the
\code{fit_ncvi()} wrapper function}

\item{pars}{List of unknown variables,
including \code{phi$mu}}

\item{differentials}{List of differentials functions, should have
'mu' and 'Sigma'}
}
\description{
Non-conjugate update for the parameters of
multivariate normal distribution.
}
