% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboVI.R
\docType{package}
\name{ncvi}
\alias{ncvi}
\title{ncvi: a package for fitting Bayesian GLMM to RNA-Seq and Ribo-Seq data
using non-conjugate variational methods.}
\description{
The package provides functions to fit a Bayesian GLMM using
a variational algorithm.
}
\section{Prep functions}{
NA
}

