% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize count table using DESeq2}
\usage{
normalize(counts, group)
}
\arguments{
\item{counts}{Data frame of counts with column \code{Geneid}}

\item{group}{Vector of group labels for each sample}
}
\description{
Normalize count table using DESeq2
}
