% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2list.R
\name{vector2list}
\alias{vector2list}
\title{Transform a vector of length \code{N * G}
to a list of vectors each of length \code{N}}
\usage{
vector2list(y, G, N)
}
\arguments{
\item{y}{The original vector}

\item{G}{Integer for the number of groups}

\item{N}{Integer for the length of each group}
}
\description{
Transform a vector of length \code{N * G}
to a list of vectors each of length \code{N}
}
